Update Diorite utils.

+++ Lobby/src/main/java/pl/arieals/lobby/chest/loot/ItemShardLoot.java
@@ -10,8 +10,8 @@ import org.bukkit.entity.Player;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
 
-import org.diorite.utils.math.DioriteRandomUtils;
-import org.diorite.utils.math.IWeightedRandomChoice;
+import org.diorite.commons.math.DioriteRandomUtils;
+import org.diorite.commons.math.IWeightedRandomChoice;
 
 import pl.arieals.globalshops.server.IGlobalShops;
 import pl.arieals.globalshops.server.IPlayerContainer;
@@ -172,7 +172,7 @@ public class ItemShardLoot implements ILoot
         }
 
         @Override
-        public int getWeight()
+        public double getWeight()
         {
             return this.weight;
         }
